Refactor of the Process Monitor code to be more resilient to errors
In https://github.com/elastic/logstash/issues/6296 we have discovered
that if the user unmount the `/dev/procfs` file the metric collection
would make logstash completely crash.

The solution is to make sure that any calls to beans are wrapped into
try/catch.

This PR do a bit of refactor to implement that behavior.